rspec: https://github.com/metanorma/metanorma-standoc/issues/479
