+++ spec/isodoc/ref_spec.rb
@@ -85,7 +85,7 @@ RSpec.describe IsoDoc do
               <formattedref format="application/x-isodoc+xml">Title 5</formattedref>
               <docidentifier type="metanorma">[5]</docidentifier>
             </bibitem>
-            <bibitem id="ISBN" type="ISBN">
+            <bibitem id="ISBN" type="book">
               <title format="text/plain">Chemicals for analytical laboratory use</title>
               <docidentifier type="ISBN">ISBN</docidentifier>
               <docidentifier type="metanorma">[1]</docidentifier>
@@ -96,7 +96,7 @@ RSpec.describe IsoDoc do
                 </organization>
               </contributor>
             </bibitem>
-            <bibitem id="ISSN" type="ISSN">
+            <bibitem id="ISSN" type="journal">
               <title format="text/plain">Instruments for analytical laboratory use</title>
               <docidentifier type="ISSN">ISSN</docidentifier>
               <docidentifier type="metanorma">[2]</docidentifier>
@@ -172,48 +172,18 @@ RSpec.describe IsoDoc do
             <title depth="1">Bibliography</title>
             <p>The following documents are referred to in the text in such a way that some or all of their content constitutes requirements of this document. For dated references, only the edition cited applies. For undated references, the latest edition of the referenced document (including any amendments) applies.</p>
             <bibitem id="ISO712" type="standard">
-              <title format="text/plain">Cereals or cereal products</title>
-              <title format="text/plain" type="main">Cereals and cereal products</title>
+              <formattedref>International Organization for Standardization. <em>Cereals and cereal products</em>.</formattedref>
               <docidentifier type="ISO">ISO 712</docidentifier>
               <docidentifier type="metanorma">[110]</docidentifier>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <name>International Organization for Standardization</name>
-                </organization>
-              </contributor>
             </bibitem>
             <bibitem id="ISO16634" type="standard">
-              <title format="text/plain" language="x">Cereals, pulses, milled cereal products, xxxx, oilseeds and animal feeding stuffs</title>
-              <title format="text/plain" language="en">Cereals, pulses, milled cereal products, oilseeds and animal feeding stuffs</title>
+              <formattedref><em>Cereals, pulses, milled cereal products, oilseeds and animal feeding stuffs</em>.</formattedref>
               <docidentifier type="ISO">ISO 16634:-- (all parts)</docidentifier>
-              <date type="published">
-                <on>--</on>
-              </date>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <abbreviation>ISO</abbreviation>
-                </organization>
-              </contributor>
               <note format="text/plain" reference="1" type="Unpublished-Status">Under preparation. (Stage at the time of publication ISO/DIS 16634)</note>
-              <extent type="part">
-                <referenceFrom>all</referenceFrom>
-              </extent>
             </bibitem>
             <bibitem id="ISO20483" type="standard">
-              <title format="text/plain">Cereals and pulses</title>
+              <formattedref>International Organization for Standardization. <em>Cereals and pulses</em>. 2013&#x2013;2014.</formattedref>
               <docidentifier type="ISO">ISO 20483:2013-2014</docidentifier>
-              <date type="published">
-                <from>2013</from>
-                <to>2014</to>
-              </date>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <name>International Organization for Standardization</name>
-                </organization>
-              </contributor>
             </bibitem>
             <bibitem id="ref1">
               <formattedref format="application/x-isodoc+xml">
@@ -221,7 +191,7 @@ RSpec.describe IsoDoc do
                 <em>Determination of the protein content in cereal and cereal products for food and animal feeding stuffs according to the Dumas combustion method</em>
                 (see
                 <link target="http://www.icc.or.at"/>
-                )</formattedref>
+                ) [ICC/167]</formattedref>
               <docidentifier type="ICC">ICC/167</docidentifier>
             </bibitem>
             <note>
@@ -232,27 +202,15 @@ RSpec.describe IsoDoc do
               <formattedref format="application/x-isodoc+xml">Title 5</formattedref>
               <docidentifier type="metanorma">[5]</docidentifier>
             </bibitem>
-            <bibitem id="ISBN" type="ISBN">
-              <title format="text/plain">Chemicals for analytical laboratory use</title>
+            <bibitem id="ISBN" type="book">
+              <formattedref><em>Chemicals for analytical laboratory use</em>. ISBN.</formattedref>
               <docidentifier type="ISBN">ISBN</docidentifier>
               <docidentifier type="metanorma">[1]</docidentifier>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <abbreviation>ISBN</abbreviation>
-                </organization>
-              </contributor>
             </bibitem>
-            <bibitem id="ISSN" type="ISSN">
-              <title format="text/plain">Instruments for analytical laboratory use</title>
+            <bibitem id="ISSN" type="journal">
+              <formattedref><em>Instruments for analytical laboratory use</em>. ISSN.</formattedref>
               <docidentifier type="ISSN">ISSN</docidentifier>
               <docidentifier type="metanorma">[2]</docidentifier>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <abbreviation>ISSN</abbreviation>
-                </organization>
-              </contributor>
             </bibitem>
             <note>
               <name>NOTE</name>
@@ -263,14 +221,8 @@ RSpec.describe IsoDoc do
               <p>This is another annotation of document ISSN.</p>
             </note>
             <bibitem id="ISO3696" type="standard">
-              <title format="text/plain">Water for analytical laboratory use</title>
+              <formattedref><em>Water for analytical laboratory use</em>.</formattedref>
               <docidentifier type="BIPM">BIPM 3696</docidentifier>
-              <contributor>
-                <role type="publisher"/>
-                <organization>
-                  <abbreviation>ISO</abbreviation>
-                </organization>
-              </contributor>
             </bibitem>
             <bibitem id="ref10">
               <formattedref format="application/x-isodoc+xml">
@@ -282,13 +234,13 @@ RSpec.describe IsoDoc do
               <docidentifier type="metanorma">[10]</docidentifier>
             </bibitem>
             <bibitem id="ref11">
-              <title>Internet Calendaring and Scheduling Core Object Specification (iCalendar)</title>
+              <formattedref><em>Internet Calendaring and Scheduling Core Object Specification (iCalendar)</em>.</formattedref>
               <docidentifier type="IETF">IETF RFC 10</docidentifier>
             </bibitem>
             <bibitem id="ref12">
               <formattedref format="application/x-isodoc+xml">CitationWorks. 2019.
                 <em>How to cite a reference</em>
-                .</formattedref>
+                . [IETF RFC 20]</formattedref>
               <docidentifier type="metanorma">[Citn]</docidentifier>
               <docidentifier type="IETF">IETF RFC 20</docidentifier>
             </bibitem>
@@ -326,31 +278,32 @@ RSpec.describe IsoDoc do
              </p>
              <p id='ISO712' class='Biblio'>
                [110]&#160; ISO 712
-               <i>Cereals and cereal products</i>
+               International Organization for Standardization.
+               <i>Cereals and cereal products</i>.
              </p>
              <p id='ISO16634' class='Biblio'>
                [2]&#160; ISO 16634:-- (all parts)
                <a class='FootnoteRef' href='#fn:1'>
                  <sup>1</sup>
                </a>
-               <i>Cereals, pulses, milled cereal products, oilseeds and animal feeding stuffs</i>
+               <i>Cereals, pulses, milled cereal products, oilseeds and animal feeding stuffs</i>.
              </p>
              <p id='ISO20483' class='Biblio'>
-               [3]&#160; ISO 20483:2013-2014
-               <i>Cereals and pulses</i>
+               [3]&#160; ISO 20483:2013-2014 International Organization for Standardization.
+               <i>Cereals and pulses</i>. 2013&#8211;2014.
              </p>
              <p id='ref1' class='Biblio'>
-               [4]&#160; ICC/167#{' '}
+               [4]&#160; ICC/167
                <span style='font-variant:small-caps;'>Standard No I.C.C 167</span>
-               .#{' '}
+               .
                <i>
                  Determination of the protein content in cereal and cereal products for
                  food and animal feeding stuffs according to the Dumas combustion
                  method
                </i>
-                (see#{' '}
+                (see
                <a href='http://www.icc.or.at'>http://www.icc.or.at</a>
-                )
+                )  [ICC/167]
              </p>
              <div class='Note'>
                <p>
@@ -360,12 +313,12 @@ RSpec.describe IsoDoc do
              </div>
              <p id='zip_ffs' class='Biblio'>[5]&#160; Title 5</p>
              <p id='ISBN' class='Biblio'>
-               [1]&#160;#{' '}
-               <i>Chemicals for analytical laboratory use</i>
+               [1]&#160;
+               <i>Chemicals for analytical laboratory use</i>. ISBN.
              </p>
              <p id='ISSN' class='Biblio'>
-               [2]&#160;#{' '}
-               <i>Instruments for analytical laboratory use</i>
+               [2]&#160;
+               <i>Instruments for analytical laboratory use</i>. ISSN.
              </p>
              <div class='Note'>
                <p>
@@ -381,29 +334,28 @@ RSpec.describe IsoDoc do
              </div>
              <p id='ISO3696' class='Biblio'>
                [8]&#160; BIPM 3696
-               <i>Water for analytical laboratory use</i>
+               <i>Water for analytical laboratory use</i>.
              </p>
              <p id='ref10' class='Biblio'>
-               [10]&#160;#{' '}
+               [10]&#160;
                <span style='font-variant:small-caps;'>Standard No I.C.C 167</span>
-               .#{' '}
+               .
                <i>
                  Determination of the protein content in cereal and cereal products for
                  food and animal feeding stuffs according to the Dumas combustion
                  method
                </i>
-                (see#{' '}
+                (see
                <a href='http://www.icc.or.at'>http://www.icc.or.at</a>
                 )
              </p>
              <p id='ref11' class='Biblio'>
                [10]&#160; IETF RFC 10
-               <i>Internet Calendaring and Scheduling Core Object Specification (iCalendar)</i>
+               <i>Internet Calendaring and Scheduling Core Object Specification (iCalendar)</i>.
              </p>
              <p id='ref12' class='Biblio'>
-               Citn&#160; IETF RFC 20CitationWorks. 2019.#{' '}
-               <i>How to cite a reference</i>
-                .
+               Citn&#160; IETF RFC 20 CitationWorks. 2019.
+               <i>How to cite a reference</i>. [IETF RFC 20]
              </p>
            </div>
            <aside id='fn:1' class='footnote'>
@@ -526,16 +478,16 @@ RSpec.describe IsoDoc do
         <references id='_normative_references' obligation='informative' normative='false' displayorder='1'>
           <title depth='1'>Bibliography</title>
           <bibitem id='ref1' type='standard'>
-            <title format='text/plain'>Cereals or cereal products</title>
+            <formattedref><em>Cereals or cereal products</em>.</formattedref>
             <docidentifier type='metanorma-ordinal'>[1]</docidentifier>
             <docidentifier>ABC</docidentifier>
           </bibitem>
           <bibitem id='ref2' type='standard' hidden='true'>
-            <title format='text/plain'>Cereals or cereal products</title>
+            <formattedref><em>Cereals or cereal products</em>.</formattedref>
             <docidentifier>ABD</docidentifier>
           </bibitem>
           <bibitem id='ref3' type='standard'>
-            <title format='text/plain'>Cereals or cereal products</title>
+            <formattedref><em>Cereals or cereal products</em>.</formattedref>
             <docidentifier type='metanorma-ordinal'>[2]</docidentifier>
             <docidentifier>ABE</docidentifier>
           </bibitem>
