+++ spec/vcr_cassettes/bipm.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: https://raw.githubusercontent.com/relaton/relaton-data-bipm/master/data/CR-03.yaml
+    uri: https://raw.githubusercontent.com/relaton/relaton-data-bipm/master/data/PV-105.yaml
     body:
       encoding: US-ASCII
       string: ''
@@ -12,7 +12,7 @@ http_interactions:
       Accept:
       - text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9
       User-Agent:
-      - Mozilla/5.0 (compatible; Konqueror/3; Linux)
+      - Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:43.0) Gecko/20100101 Firefox/43.0
       Accept-Charset:
       - ISO-8859-1,utf-8;q=0.7,*;q=0.7
       Accept-Language:
@@ -49,33 +49,33 @@ http_interactions:
       Content-Type:
       - text/plain; charset=utf-8
       X-Github-Request-Id:
-      - 91A2:5B02:370D5D:4A8BBF:625D6174
+      - C072:4508:82751:1061FF:6262C8F6
       Accept-Ranges:
       - bytes
       Date:
-      - Mon, 18 Apr 2022 13:02:44 GMT
+      - Fri, 22 Apr 2022 15:25:42 GMT
       Via:
       - 1.1 varnish
       X-Served-By:
-      - cache-mel11260-MEL
+      - cache-mel11243-MEL
       X-Cache:
       - MISS
       X-Cache-Hits:
       - '0'
       X-Timer:
-      - S1650286964.203098,VS0,VE308
+      - S1650641142.454322,VS0,VE332
       Vary:
       - Authorization,Accept-Encoding,Origin
       Access-Control-Allow-Origin:
       - "*"
       X-Fastly-Request-Id:
-      - ec8c262bd4bb1a72012c338aa69f83b48c1beb92
+      - 8a1fcc4f0ee04ae1ae9f37d93e8d7f10c2eeaa4e
       Expires:
-      - Mon, 18 Apr 2022 13:07:44 GMT
+      - Fri, 22 Apr 2022 15:30:42 GMT
       Source-Age:
       - '0'
     body:
       encoding: UTF-8
       string: '404: Not Found'
-  recorded_at: Mon, 18 Apr 2022 13:02:44 GMT
+  recorded_at: Fri, 22 Apr 2022 15:25:42 GMT
 recorded_with: VCR 6.1.0
